[MIX] - [Cubemap fetching rework; Rendering modes fix;  ]

+++ src/project/hooks/useSerializer.js
@@ -55,7 +55,7 @@ export default function useSerializer(engine, setAlert, settings, id, quickAcces
 
                         await Promise.all(all.map(a => {
                             return new Promise(async (resolve1) => {
-                                if (a && !engine.entities.find(e => e.id === a.id))
+                                if (a && a.data && !engine.entities.find(e => e.id === a.id))
                                     resolve1(await fileSystem.deleteFile(fileSystem.path + '\\logic\\' + a.data.id + '.entity', true))
                                 else
                                     resolve1()
